adding production dockerfile and adding some env variables to set up the production database connection

+++ api/main.py
@@ -2,7 +2,10 @@ from dotenv import load_dotenv
 # Load environment variables from .env file
 load_dotenv()
 
+
 import os
+
+
 
 from fastapi import FastAPI
 from fastapi.responses import RedirectResponse
@@ -12,18 +15,22 @@ from api.services.database import sqlite_file_name, create_db_and_tables
 from api.services.dummy_data import create_dummy_data
 from api.routers import users, events, ipfs
 
-# --------------------------------------
-# CREATING DUMMY DATABASE AND DATA
-# ADD dummy data in sqlite database (Only Dev Mode)----
-try:
-    os.remove(sqlite_file_name)  # START FROM SCRATCH
-    print("DB was removed")
-except:
+
+if os.environ["ENVIRONMENT"] == "PROD":
     pass
-create_db_and_tables()
-print("Tables were created")
-create_dummy_data()
-# -------------------------------------
+else:
+    # --------------------------------------
+    # CREATING DUMMY DATABASE AND DATA
+    # ADD dummy data in sqlite database (Only Dev Mode)----
+    try:
+        os.remove(sqlite_file_name)  # START FROM SCRATCH
+        print("DB was removed")
+    except:
+        pass
+    create_db_and_tables()
+    print("Tables were created")
+    create_dummy_data()
+    # -------------------------------------
 
 description = """
     This api is created to serve the neccesities
